ADD LOGIC TO CHECK THAT DIMENSION < 100
Edited the existing function to create the grid, so that it now checks
whether dimension is less than 100. Otherwise, it will alert the user to
please choose a smaller number.

Also added 2 buttons between the grid and the headline:
1. the first button should allow the user to change the grid size
2. the second button should allow the user to clear the canvas (reset
the class so that the background color goes back to default)

The logic has not yet been added, just the buttons.