CREATE FUNCTION RESIZE GRID
Created function resize grid which should ask the user to give the new
grid size in a prompt, and store the result in the dimension variable.
Then re-trigger grid creation and reapply eventlisteners to restore the
functionality. This is still very much work in progress and has quite a
few problems that need investigating (see comments).