CREATE RUNETCHASKETCH FUNCTION
I created a new function runEtchASketch which encompasses the
adding of eventlisteners to the cells and applying the mousover
function on them. This allows me to add eventlisteners and establish
the same "etch a sketch" functionality that was present on page load
after a new grid was created (by calling the function).